refactor(graph): extract sigma global route shell
